new dts file
Origin file for varsomam33 was placed in kernel source. But in yocto environment layer for kernel is more sutable. Now machine varsomamm33c has 2 dts - one for DK the socond for target PLC. Origin file is stil in the kernel and not supposed to be modificated. It will be used to run the Linux on DK.